requirements.txt
flask==3.1.2
tensorflow==2.20.0
numpy==2.2.6
python-dotenv==1.2.1